+++ packages/dmark/scripts/nodejs/test.js
@@ -2,7 +2,7 @@ const { runQueue, cmdTask } = require('../../src/utils');
 
 const watch = process.argv.includes('--watch');
 const coverage = process.argv.includes('--coverage');
-const jestCmd = 'jest --config=./config/jest.config.js --forceExit --detectOpenHandles'.split(' ');
+const jestCmd = 'jest --config=./config/jest.config.js --forceExit --detectOpenHandles';
 
 jestCmd.push(watch ? '--watch' : '--watchAll=false');